Added Mike Fischer as a contributor.
($VERSION): Incremented to 1.10.

(receive_response_3): Added optional "dont_block" argument.  If this
is present and non-zero, pass MSG_DONTWAIT to the recv() call.
MSG_DONTWAIT is wrapped in an eval, to avoid breaking the code on
systems that don't have the flag.

(request_response_5): Pass dont_block=1 to receive_response_3.
According to Mike Fisher, Linux sometimes blocks on recv() even though
a select() for readability has returned, for example when a checksum
fails.